// src/constants/tailorStatus.enum.ts
export enum TailorAssignmentStatus {
  NOT_ASSIGNED = 'NOT_ASSIGNED',
  ASSIGNED = 'ASSIGNED',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  UNASSIGNED = 'UNASSIGNED',
}

export enum TailorReturnStatus {
  NOT_COMPLETED = 'NOT_COMPLETED',
  PARTIALLY_COMPLETED = 'PARTIALLY_COMPLETED',
  COMPLETED = 'COMPLETED',
}